move generic op types into separate header
